+++ server/libs/modules/components/llm/amazon-bedrock/src/main/java/com/bytechef/component/amazon/bedrock/action/AmazonBedrockAnthropic3ChatAction.java
@@ -98,9 +98,9 @@ public class AmazonBedrockAnthropic3ChatAction {
         @Override
         public ChatOptions createChatOptions(Parameters inputParameters) {
             return Anthropic3ChatOptions.builder()
-                .withTemperature(inputParameters.getFloat(TEMPERATURE))
+                .withTemperature(inputParameters.getDouble(TEMPERATURE))
                 .withMaxTokens(inputParameters.getInteger(MAX_TOKENS))
-                .withTopP(inputParameters.getFloat(TOP_P))
+                .withTopP(inputParameters.getDouble(TOP_P))
                 .withStopSequences(inputParameters.getList(STOP, new TypeReference<>() {}))
                 .withTopK(inputParameters.getInteger(TOP_K))
                 .withAnthropicVersion("bedrock-2023-05-31")
